% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geowaterdata.R
\docType{data}
\name{Area_data}
\alias{Area_data}
\title{Area_data}
\format{
A shape file with the polygon of the study area:
\describe{
\item{x}{The min and max latitud}
\item{y}{The min and max longitud}
}
}
\usage{
Area_data
}
\description{
Area_data
}
\keyword{datasets}
