% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geowaterdata.R
\docType{data}
\name{Balance}
\alias{Balance}
\title{Balance}
\format{
A csv file that contains identification, logtitud, latitud, type of source and the hydrochemical composition of the water samples colected in the Magdalena Middle Valley, Colombia:
\describe{
\item{ID}{Identification of each water sample}
\item{long}{Longitud of the point}
\item{lat}{Latitud of the point}
}
}
\usage{
Balance
}
\description{
Balance
}
\keyword{datasets}
