% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geowaterdata.R
\docType{data}
\name{Dataclust}
\alias{Dataclust}
\title{Dataclust}
\format{
A csv file that contains identification, logtitud, latitud, type of source, cluster clasification using \code{\link{waterclust}} function and, the hydrochemical composition of the water samples colected in the Magdalena Middle Valley, Colombia:
\describe{
\item{ID}{Identification of each water sample}
\item{long}{Longitud of the point}
\item{lat}{Latitud of the point}
}
}
\usage{
Dataclust
}
\description{
Dataclust
}
\keyword{datasets}
