% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geowaterdata.R
\docType{data}
\name{geological_units}
\alias{geological_units}
\title{Geological Units}
\format{
A shape file with the geology of the study area (Magdalena Middle Valley, Colombia):
\describe{
\item{x}{The min and max latitud}
\item{y}{The min and max longitud}
\item{SimboloUC}{The geological descrition of each unit}
}
}
\source{
\url{https://www.datos.gov.co/Minas-y-Energ-a/Mapa-Geol-gico-de-Colombia-cronoestratigraf-a/5fee-j9jk}
}
\usage{
geological_units
}
\description{
Geological Units
}
\keyword{datasets}
