% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAcoda.R
\name{PCAcoda}
\alias{PCAcoda}
\title{Principal Analysis Components using the Compositional Data (CoDa).}
\usage{
PCAcoda(Dataclust, comp1 = 1, comp2 = 2)
}
\arguments{
\item{Dataclust}{is a matrix that contains the hydrochemical composition of water samples and the assigned cluster using the \code{\link{waterclut}} function. Sample name, source, long, lat, Mg, Ca, Na,K,HCO3, Cl,SO4,NO3,NO2,Fe. All concentrations are in meq/l.}

\item{comp1}{is a numeric value indicating the number of the component to be plotted on the x-axys.}

\item{comp2}{is a numeric value indicating the number of the component to be plotted on the y-axys.}
}
\value{

}
\description{
Principal Analysis COmponents using the Compositional Data (CoDa) approach for the data treatment and a clr transformation (\cite{Aitchison1982a})
}
\examples{
PCAcoda(Dataclust, comp1 = 1, comp2 = 2)

}
\author{
Adriana Piña \href{mailto:appinaf@unal.edu.co}{appinaf@unal.edu.co}\cr
David Zamora \href{mailto:dazamoraa@unal.edu.co}{dazamoraa@unal.edu.co} \cr
}
