% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterclust.R
\name{waterclust}
\alias{waterclust}
\title{Title}
\usage{
waterclust(Data, height, typ)
}
\arguments{
\item{Data}{is dataframe that contains the hydrochemical composition of water samples. Sample name, source, long, lat, Mg, Ca, Na,K,HCO3, Cl,SO4,NO3,NO2,Fe. All concentrations are in meq/l.}

\item{height}{is an integer that defines the height of the treecut to define clusters}

\item{typ}{is a double between 0 and 1 which defines the analysis type. 1 for the samples clustering and 2 for the chemical compounds clustering.}
}
\description{
This function allows the user to define clusters according to the hydrochemical characteristics of water samples and to build dendograms, using the Compositional Data (CoDa) approach for the data treatment and a clr transformation (\cite{Aitchison1982a}).
}
\examples{
Dataclust <- waterclust(Data, height = 35, typ = 2)
}
\author{
Adriana Piña \href{mailto:appinaf@unal.edu.co}{appinaf@unal.edu.co}\cr
David Zamora \href{mailto:dazamoraa@unal.edu.co}{dazamoraa@unal.edu.co} \cr
}
