% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geowaterdata.R
\docType{data}
\name{Area_data}
\alias{Area_data}
\title{Area.data}
\format{
A data frame with X rows and Y variables:
\describe{
\item{X}{description 1}
\item{Y}{description 2}
...
}
}
\source{
\url{http://www.XXXX.info/}
}
\usage{
Area_data
}
\description{
Area.data
}
\keyword{datasets}
