% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geowaterdata.R
\docType{data}
\name{faults}
\alias{faults}
\title{faults}
\format{
A shape file with the faults of the study area (Magdalena Middle Valley, Colombia):
\describe{
\item{x}{The min and max latitud}
\item{y}{The min and max longitud}
\item{SimboloUC}{The name of the fault}
\item{CRS arguments} {The projection of the shape file}
}
}
\source{
\url{https://www.datos.gov.co/dataset/Fallas/8gux-3y6i}
}
\usage{
faults
}
\description{
faults
}
\keyword{datasets}
