% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAcoda.R
\name{PCAcoda}
\alias{PCAcoda}
\title{Principal Analysis Components using the Compositional Data (CoDA) approach.}
\usage{
PCAcoda(Dataclust, comp1, comp2)
}
\arguments{
\item{Dataclust}{is a matrix that contains the hydrochemical composition of water samples and the assigned cluster using the \code{\link{waterclust}} function.
Titles must be as follows: ID, long, lat, source, Mg, Ca, Na, K, HCO3, Cl,SO4, NO3, NO2, Fe. All concentrations are in meq/l.
Aditional chemical compounds must be added in columns after the Fe column concentration.}

\item{comp1}{is a numeric value indicating the number of the component to be plotted on the x-axys.}

\item{comp2}{is a numeric value indicating the number of the component to be plotted on the y-axys.}
}
\value{
returns an interactive compositional biplot with the selected compositions, the samples description and the assigned cluster from the \code{\link{waterclust}} function;
a summary of the PCA analysis with the Standard deviation, the Proportion of Variance and the Cumulative Proportion for each composition;
a list with class "princomp" (see \code{\link{compositions}} documentation)

a list that comprises a figure and two dataframes. The figure is an interactive compositional biplot with the selected compositions, the samples description and the assigned cluster from the \code{\link{waterclust}} function;
The first list, is the summary of the PCA analysis with the Standard deviation, the Proportion of Variance and the Cumulative Proportionfor each composition.
The second one is a list with class "princomp" (see \code{\link{compositions}} documentation).
}
\description{
Principal Analysis Components using the Compositional Data (CoDA) approach for the data treatment and the Centered log ratio - clr transformation \cite{(Aitchison, 1982)}.
}
\examples{
PCAcoda(Dataclust, comp1 = 1, comp2 = 2)
}
\references{
Aitchison, J. (1982). The Statistical Analysis of Compositional Data. Journal of the Royal Statistical Society. Series B (Methodological), 44(2), 139–177.
}
\author{
Adriana Piña \href{mailto:appinaf@unal.edu.co}{appinaf@unal.edu.co}\cr
David Zamora \href{mailto:dazamoraa@unal.edu.co}{dazamoraa@unal.edu.co} \cr
}
